requirements.txt
fastapi
uvicorn[standard]
jinja2
qrcode
pillow
python-multipart
pytailwindcss
rembg[cpu]
opencv-python-headless
numpy
fpdf2
prophet
yfinance
pandas
scikit-learn
